Attempt at Rendering batched and other stuff and gui rework

+++ HeliosEditor/Shaders/DynamicGizmosVertexShader.hlsl
@@ -6,11 +6,11 @@ cbuffer CBuf
 struct VSOut
 {
     float4 color : Color;
-    float id : Id;
+    int id : Id;
     float4 position : SV_Position;
 };
 
-VSOut main( float3 pos : Position, float4 col : Color, float id : Id)
+VSOut main( float3 pos : Position, float4 col : Color, int id : Id)
 {
     VSOut vout;
     vout.position = mul(float4(pos, 1.0f), view_matrix);
