+++ HeliosEditor/Shaders/LineVertexShader.hlsl
@@ -8,11 +8,11 @@ struct VSOut
     float4 col : Color;
     float thickness : Thickness;
     int mode : Mode;
-    float id : Id;
+    int id : Id;
     float4 pos : SV_Position;
 };
 
-VSOut main(float3 pos : Position, float4 col : Color, int mode : Mode, float thickness : Thickness, float id : Id)
+VSOut main(float3 pos : Position, float4 col : Color, int mode : Mode, float thickness : Thickness, int id : Id)
 {
     VSOut vout;
     vout.pos = mul(float4(pos, 1.0f), viewProjection);
