+++ HeliosEditor/Shaders/QuadGizmoVertexShader.hlsl
@@ -1,7 +1,7 @@
 struct VSOut
 {
     float4 color : Color;
-    float data : Data;
+    int data : Data;
     float4 position : SV_Position;
 };
 
@@ -10,7 +10,7 @@ cbuffer Cbuf
     matrix viewMatrix;
 }
 
-VSOut main(float2 pos : Position, matrix world : World, float4 color : Color, float data : Data)
+VSOut main(float2 pos : Position, matrix world : World, float4 color : Color, int data : Data)
 {
     VSOut vso;
     vso.position = mul(float4(pos.x, pos.y, 0.0f, 1.0f), mul(world, viewMatrix));
