+++ HeliosEditor/Shaders/StandardGizmosVertexShader.hlsl
@@ -1,7 +1,7 @@
 struct VSOut
 {
     float4 color : Color;
-    float id : Id;
+    int id : Id;
     float4 position : SV_Position;
 };
 
@@ -10,7 +10,7 @@ cbuffer Cbuf
     matrix viewMatrix;
 }
 
-VSOut main(float3 position : Position, matrix transform : Transform, float4 color : Color, float id : Id)
+VSOut main(float3 position : Position, matrix transform : Transform, float4 color : Color, int id : Id)
 {
     VSOut vso;
     //vso.position = float4(position, 1.0f);
